Use a more specific default for USER_ID_FIELD
The 'pk' field in the django ORM has a fluid meaning.  App developers
could decide to use an 'email' field as the primary key for their user
model.  This might lead to unintended vulnerabilities if the field's
value is allowed to change after it is originally set (as might be the
case with a 'email' or 'username' field).  Therefore, we make 'id' the
default user id field since there is a stronger implication that this
field's value will never change after a user record is created.